Make it possible to complete all types of notification
