+++ src/main/kotlin/no/nav/syfo/service/VarselBusService.kt
@@ -1,13 +1,16 @@
 package no.nav.syfo.service
 
+import no.nav.syfo.ARBEIDSGIVERNOTIFIKASJON_OPPFOLGING_MERKELAPP
 import no.nav.syfo.kafka.consumers.varselbus.domain.ArbeidstakerHendelse
 import no.nav.syfo.kafka.consumers.varselbus.domain.EsyfovarselHendelse
 import no.nav.syfo.kafka.consumers.varselbus.domain.HendelseType.*
 import no.nav.syfo.kafka.consumers.varselbus.domain.NarmesteLederHendelse
+import no.nav.syfo.kafka.consumers.varselbus.domain.toVarselData
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
 class VarselBusService(
+    val senderFacade: SenderFacade,
     val motebehovVarselService: MotebehovVarselService,
     val oppfolgingsplanVarselService: OppfolgingsplanVarselService,
     val dialogmoteInnkallingVarselService: DialogmoteInnkallingVarselService,
@@ -18,27 +21,45 @@ class VarselBusService(
         varselHendelse: EsyfovarselHendelse
     ) {
         log.info("Behandler varsel av type ${varselHendelse.type}")
-        when (varselHendelse.type) {
-            NL_OPPFOLGINGSPLAN_SENDT_TIL_GODKJENNING -> oppfolgingsplanVarselService.sendEllerFerdigstillVarselTilNarmesteLeder(varselHendelse.toNarmestelederHendelse())
-            SM_OPPFOLGINGSPLAN_SENDT_TIL_GODKJENNING -> oppfolgingsplanVarselService.sendEllerFerdigstillVarselTilArbeidstaker(varselHendelse.toArbeidstakerHendelse())
+        if (varselHendelse.skalFerdigstilles()) {
+            ferdigstillVarsel(varselHendelse)
+        } else {
+            when (varselHendelse.type) {
+                NL_OPPFOLGINGSPLAN_SENDT_TIL_GODKJENNING -> oppfolgingsplanVarselService.sendVarselTilNarmesteLeder(varselHendelse.toNarmestelederHendelse())
+                SM_OPPFOLGINGSPLAN_SENDT_TIL_GODKJENNING -> oppfolgingsplanVarselService.sendVarselTilArbeidstaker(varselHendelse.toArbeidstakerHendelse())
 
-            NL_DIALOGMOTE_SVAR_MOTEBEHOV -> motebehovVarselService.sendVarselTilNarmesteLeder(varselHendelse.toNarmestelederHendelse())
-            SM_DIALOGMOTE_SVAR_MOTEBEHOV -> motebehovVarselService.sendVarselTilArbeidstaker(varselHendelse.toArbeidstakerHendelse())
+                NL_DIALOGMOTE_SVAR_MOTEBEHOV -> motebehovVarselService.sendVarselTilNarmesteLeder(varselHendelse.toNarmestelederHendelse())
+                SM_DIALOGMOTE_SVAR_MOTEBEHOV -> motebehovVarselService.sendVarselTilArbeidstaker(varselHendelse.toArbeidstakerHendelse())
 
-            NL_DIALOGMOTE_INNKALT,
-            NL_DIALOGMOTE_AVLYST,
-            NL_DIALOGMOTE_REFERAT,
-            NL_DIALOGMOTE_NYTT_TID_STED -> dialogmoteInnkallingVarselService.sendVarselTilNarmesteLeder(varselHendelse.toNarmestelederHendelse())
+                NL_DIALOGMOTE_INNKALT,
+                NL_DIALOGMOTE_AVLYST,
+                NL_DIALOGMOTE_REFERAT,
+                NL_DIALOGMOTE_NYTT_TID_STED -> dialogmoteInnkallingVarselService.sendVarselTilNarmesteLeder(varselHendelse.toNarmestelederHendelse())
 
-            SM_DIALOGMOTE_INNKALT,
-            SM_DIALOGMOTE_AVLYST,
-            SM_DIALOGMOTE_REFERAT,
-            SM_DIALOGMOTE_NYTT_TID_STED,
-            SM_DIALOGMOTE_LEST -> dialogmoteInnkallingVarselService.sendVarselTilArbeidstaker(varselHendelse.toArbeidstakerHendelse())
+                SM_DIALOGMOTE_INNKALT,
+                SM_DIALOGMOTE_AVLYST,
+                SM_DIALOGMOTE_REFERAT,
+                SM_DIALOGMOTE_NYTT_TID_STED,
+                SM_DIALOGMOTE_LEST -> dialogmoteInnkallingVarselService.sendVarselTilArbeidstaker(varselHendelse.toArbeidstakerHendelse())
 
-            else -> {
-                log.warn("Klarte ikke mappe varsel av type ${varselHendelse.type} ved behandling forsøk")
+                else -> {
+                    log.warn("Klarte ikke mappe varsel av type ${varselHendelse.type} ved behandling forsøk")
+                }
             }
+        }
+    }
+
+    fun ferdigstillVarsel(varselHendelse: EsyfovarselHendelse) {
+        if (varselHendelse.isArbeidstakerHendelse()) {
+            senderFacade.ferdigstillBrukernotifkasjonVarsler(varselHendelse.toArbeidstakerHendelse())
+        } else {
+            senderFacade.ferdigstillDineSykmeldteVarsler(
+                varselHendelse.toNarmestelederHendelse()
+            )
+            senderFacade.ferdigstillArbeidsgiverNotifikasjoner(
+                varselHendelse.toNarmestelederHendelse(),
+                ARBEIDSGIVERNOTIFIKASJON_OPPFOLGING_MERKELAPP
+            )
         }
     }
 
@@ -66,4 +87,15 @@ class VarselBusService(
             this
         } else throw IllegalArgumentException("Wrong type of EsyfovarselHendelse, should be of type ArbeidstakerHendelse")
     }
+
+    private fun EsyfovarselHendelse.isArbeidstakerHendelse(): Boolean {
+        return type.name.startsWith("SM")
+    }
+
+
+    private fun EsyfovarselHendelse.skalFerdigstilles() =
+        data?.let { data ->
+            val varseldata = data.toVarselData()
+            varseldata.status?.ferdigstilt
+        } ?: false
 }
